Refactor search param pages into dynamic client boundaries
